requirements.txt
lxml==5.2.2
requests==2.32.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
mysql-connector-python==9.0.0